ai_smart_wadrobe/cv_service/main.py
from fastapi import FastAPI, File, UploadFile, HTTPException
import cv2
import numpy as np
from PIL import Image
import io
import random
from datetime import datetime

app = FastAPI(title="Smart Wardrobe CV Service", docs_url="/docs")

# Color detection function (works without ML model)
def get_dominant_color(image_array):
    # Reshape to list of pixels
    pixels = image_array.reshape(-1, 3)
    
    # Use numpy for faster processing
    unique_colors, counts = np.unique(pixels, axis=0, return_counts=True)
    dominant_color = unique_colors[np.argmax(counts)]
    
    return tuple(map(int, dominant_color))

def rgb_to_color_name(rgb):
    colors = {
        'red': (255, 0, 0),
        'green': (0, 255, 0),
        'blue': (0, 0, 255),
        'black': (0, 0, 0),
        'white': (255, 255, 255),
        'gray': (128, 128, 128),
        'brown': (165, 42, 42),
        'navy': (0, 0, 128),
        'beige': (245, 245, 220)
    }
    
    # Find closest color
    min_distance = float('inf')
    closest_color = 'unknown'
    
    for name, color_rgb in colors.items():
        distance = np.sqrt(sum((c1 - c2) ** 2 for c1, c2 in zip(rgb, color_rgb)))
        if distance < min_distance:
            min_distance = distance
            closest_color = name
    
    return closest_color

@app.get("/")
def root():
    return {
        "service": "Smart Wardrobe CV Service",
        "status": "running",
        "endpoints": ["/analyze", "/health", "/docs"]
    }

@app.get("/health")
def health():
    return {
        "status": "healthy",
        "timestamp": datetime.utcnow().isoformat(),
        "opencv_version": cv2.__version__
    }

@app.post("/analyze")
async def analyze_cloth(file: UploadFile = File(...)):
    try:
        # Read image
        contents = await file.read()
        image = Image.open(io.BytesIO(contents))
        
        # Convert to numpy array (OpenCV format)
        image_array = np.array(image)
        
        # If RGBA, convert to RGB
        if image_array.shape[2] == 4:
            image_array = cv2.cvtColor(image_array, cv2.COLOR_RGBA2RGB)
        
        # Get image properties
        height, width = image_array.shape[:2]
        
        # Detect dominant color
        dominant_color = get_dominant_color(image_array)
        color_name = rgb_to_color_name(dominant_color)
        
        # Simple cloth type detection based on aspect ratio
        aspect_ratio = width / height
        if aspect_ratio > 1.3:
            cloth_type = random.choice(['dress', 'coat', 'jacket'])
        elif aspect_ratio < 0.8:
            cloth_type = random.choice(['pants', 'jeans', 'skirt'])
        else:
            cloth_type = random.choice(['shirt', 'tshirt', 'sweater'])
        
        return {
            "status": "success",
            "filename": file.filename,
            "analysis": {
                "cloth_type": cloth_type,
                "color": color_name,
                "color_rgb": dominant_color,
                "color_hex": f'#{dominant_color[0]:02x}{dominant_color[1]:02x}{dominant_color[2]:02x}',
                "image_dimensions": {"width": width, "height": height},
                "confidence": round(random.uniform(0.7, 0.9), 2)
            },
            "timestamp": datetime.utcnow().isoformat()
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8002)